fix build: move FlowViewPage and FlowVisualizer into frontend
